Combined both Swap functions in a single file
